refactor: 분석 페이지 아키텍처 개선 (Server/Client) #38
- app/analysis/page.tsx를 서버 컴포넌트로 전환
- 상태 관리 및 UI 로직을 AnalysisClient 컴포넌트로 추출
